// ts_axios/src/axios/types.tsx
import AxiosInterceptorManager from "./AxiosInterceptorManager";

export type Methods =
  | "get"
  | "GET"
  | "post"
  | "put"
  | "delete"
  | "options"
  | "OPTIONS"
  | "POST"
  | "PUT"
  | "DELETE";

export interface AxiosRequestConfig {
  url?: string;
  method?: Methods;
  params?: any;
  headers?: any;
  data?: Record<string, any>;
  timeout?: number;
  transformRequest?: (data: any, headers: any) => any;
  transformResponse?: (data: any) => any;
  cancelToken?:any
}

/* 
    Axios.prototype.request这个方法
    Promise的泛型T代表此promise变成成功态之后resolve的值 resolve(Value)
*/
export interface AxiosInstance {
  <T = any>(config: AxiosRequestConfig): Promise<AxiosResponse<T>>;
  interceptors: {
    request: AxiosInterceptorManager<AxiosRequestConfig>;
    response: AxiosInterceptorManager<AxiosResponse>;
  };
  cancelToken: any;
  isCancel: any;
}

/* 泛型T代表响应体的类型 */
export interface AxiosResponse<T = any> {
  data: T;
  status: number;
  statusText: string;
  /* 
        Record<string,any> 等价于
        interface PlainObject{
            [name:string]:any
        }
    */
  headers: Record<string, any>;
  config: AxiosRequestConfig;
  request: XMLHttpRequest;
}